requirements.txt
asgiref==3.7.2
Django==4.2.4
django-comment-migrate==0.1.7
django-cors-headers==4.2.0
django-filter==23.2
django-ranged-response==0.2.0
django-restql==0.15.3
django-simple-captcha==0.5.18
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pyPEG2==2.15.2
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1